Update README.md
Hi there! I'm Karna, a motivated and detail-oriented developer with a strong foundation in computer science engineering and a passion for crafting innovative solutions. My journey in technology began with exploring software engineering principles, and it has grown to include expertise in Python, web development, and an emerging focus on Artificial Intelligence (AI) and Machine Learning (ML).

Whether it's designing scalable applications, debugging complex issues, or learning the latest in tech, I thrive on challenges and aim to contribute meaningful solutions to the digital world.